Conflicts:
	src/main/java/au/moodflip/comm/controller/CommentController.java
	src/main/java/au/moodflip/comm/controller/NotificationController.java
	src/main/java/au/moodflip/comm/controller/PrivateMessageController.java
	src/main/java/au/moodflip/comm/dao/NotificationDao.java
	src/main/java/au/moodflip/comm/dao/NotificationDaoImpl.java
	src/main/java/au/moodflip/comm/model/Topic.java
	src/main/java/au/moodflip/comm/model/TopicComment.java
	src/main/java/au/moodflip/comm/service/NotificationService.java
	src/main/java/au/moodflip/comm/service/NotificationServiceImpl.java

+++ src/main/java/au/moodflip/userpage/model/Question.java
@@ -8,7 +8,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-
+//
 @Entity
 @Table(name = "question")
 public class Question {
@@ -16,41 +16,34 @@ public class Question {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int id;
+	private long id;
 	
 	@Column(name = "text")
 	private String text;
 	
-
 	@ManyToOne
 	@JoinColumn(name="category_id") 
 	private Category category;
 
-
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
-
 
 	public String getText() {
 		return text;
 	}
 
-
 	public void setText(String text) {
 		this.text = text;
 	}
 
-
 	public Category getCategory() {
 		return category;
 	}
-
 
 	public void setCategory(Category category) {
 		this.category = category;
